+++ src/FormConfig.php
@@ -72,7 +72,7 @@ class FormConfig
                     ],
                     ChoiceRadioField::class    => [
                         'edit'    => $tpl_dir.'/edit/choice-radio.php',
-                        'display' => $tpl_dir.'/display/text.php'
+                        'display' => $tpl_dir.'/display/choice-radio.php'
                     ],
                     ChoiceOrOtherField::class  => [
                         'edit'    => $tpl_dir.'/edit/choice-or-other.php',